Update README.md
Adding extra documentation to include my repo which is a simple integration between Ionic&Angular with Meteor. This repo goes through the specific tasks to accomplish this. It is a very-easy to understand repo with a very minimal amount of code in it. It would be very useful for a beginner as it shows the steps I had to take in order to integrate these technologies together.

I've also included https://github.com/netanelgilad/meteor-ionic-example's repo as it is more for an advanced user but still very helpful.